+++ src/ModalManager.js
@@ -15,6 +15,7 @@ import {
 import { ConfirmationModal } from '@folio/stripes/components';
 
 import CheckinNoteModal from './components/CheckinNoteModal';
+import ClaimedReturnedModal from './components/ClaimedReturnedModal';
 import MultipieceModal from './components/MultipieceModal';
 import { statuses } from './consts';
 import { getFullName } from './util';
@@ -26,6 +27,7 @@ class ModalManager extends React.Component {
     intl: PropTypes.object,
     checkedinItem: PropTypes.object.isRequired,
     checkinNotesMode: PropTypes.bool,
+    claimedReturnedHandler: PropTypes.func,
     onDone: PropTypes.func.isRequired,
     onCancel: PropTypes.func.isRequired,
   };
@@ -36,6 +38,14 @@ class ModalManager extends React.Component {
     const { checkedinItem, checkinNotesMode } = props;
     this.state = { checkedinItem, checkinNotesMode };
     this.steps = [
+      {
+        validate: this.shouldClaimedReturnedModalBeShown,
+        exec: () => this.setState({ showClaimedReturnedModal: true }),
+      },
+      {
+        validate: this.shouldWithdrawnModalBeShown,
+        exec: () => this.setState({ showWithdrawnModal: true }),
+      },
       {
         validate: this.shouldDeclaredLostModalBeShown,
         exec: () => this.setState({ showDeclaredLostModal: true }),
@@ -74,9 +84,19 @@ class ModalManager extends React.Component {
     return this.props.onDone();
   }
 
+  shouldClaimedReturnedModalBeShown = () => {
+    const { checkedinItem } = this.state;
+    return checkedinItem?.status?.name === statuses.CLAIMED_RETURNED;
+  }
+
+  shouldWithdrawnModalBeShown = () => {
+    const { checkedinItem } = this.state;
+    return checkedinItem?.status?.name === statuses.WITHDRAWN;
+  }
+
   shouldDeclaredLostModalBeShown = () => {
     const { checkedinItem } = this.state;
-    return get(checkedinItem, 'status.name') === statuses.DECLARED_LOST;
+    return checkedinItem?.status?.name === statuses.DECLARED_LOST;
   }
 
   shouldCheckinNoteModalBeShown = () => {
@@ -87,7 +107,7 @@ class ModalManager extends React.Component {
 
   shouldMissingModalBeShown = () => {
     const { checkedinItem } = this.state;
-    return get(checkedinItem, 'status.name') === statuses.MISSING;
+    return checkedinItem?.status?.name === statuses.MISSING;
   }
 
   shouldMultipieceModalBeShown = () => {
@@ -107,16 +127,25 @@ class ModalManager extends React.Component {
     );
   }
 
+  confirmClaimedReturnedModal = (resolution) => {
+    this.props.claimedReturnedHandler(resolution);
+    this.setState({ showClaimedReturnedModal: false }, () => this.execSteps(1));
+  }
+
+  confirmWithdrawnModal = () => {
+    this.setState({ showWithdrawnModal: false }, () => this.execSteps(2));
+  }
+
   confirmDeclareLostModal = () => {
-    this.setState({ showDeclaredLostModal: false }, () => this.execSteps(1));
+    this.setState({ showDeclaredLostModal: false }, () => this.execSteps(3));
   }
 
   confirmMultipieceModal = () => {
-    this.setState({ showMultipieceModal: false }, () => this.execSteps(2));
+    this.setState({ showMultipieceModal: false }, () => this.execSteps(4));
   }
 
   confirmMissingModal = () => {
-    this.setState({ showMissingModal: false }, () => this.execSteps(3));
+    this.setState({ showMissingModal: false }, () => this.execSteps(5));
   }
 
   confirmCheckinNoteModal = () => {
@@ -126,44 +155,93 @@ class ModalManager extends React.Component {
   onCancel = () => {
     this.setState({
       showCheckinNoteModal: false,
+      showClaimedReturnedModal: false,
       checkinNotesMode: false,
       showMissingModal: false,
       showMultipieceModal: false,
       showDeclaredLostModal: false,
+      showWithdrawnModal: false,
     });
 
     this.props.onCancel();
   }
 
-  renderDeclaredLostModal() {
+  renderClaimedReturnedModal() {
     const {
       checkedinItem,
-      showDeclaredLostModal,
+      showClaimedReturnedModal,
     } = this.state;
+
+    return (
+      <ClaimedReturnedModal
+        item={checkedinItem}
+        open={showClaimedReturnedModal}
+        onCancel={this.onCancel}
+        onConfirm={this.confirmClaimedReturnedModal}
+      />
+    );
+  }
+
+  getItemValues = (item) => {
     const {
       barcode,
       title,
       materialType,
-    } = checkedinItem;
+    } = item;
+
+    return {
+      title,
+      barcode,
+      materialType: upperFirst(materialType?.name ?? ''),
+    };
+  }
+
+  renderConfirmStatusModal = (isOpen, status, messageId, onConfirm) => {
+    const { checkedinItem } = this.state;
 
     return (
       <ConfirmationModal
-        id="test-declared-lost-modal"
-        open={showDeclaredLostModal}
+        id={`test-${status}-modal`}
+        open={isOpen}
         item={checkedinItem}
-        heading={<FormattedMessage id="ui-checkin.declaredLostModal.heading" />}
+        heading={<FormattedMessage id={`ui-checkin.${status}Modal.heading`} />}
         message={<SafeHTMLMessage
-          id="ui-checkin.declaredLostModal.message"
-          values={{
-            title,
-            barcode,
-            materialType: upperFirst(materialType?.name ?? ''),
-          }}
+          id={messageId}
+          values={this.getItemValues(checkedinItem)}
         />}
-        onConfirm={this.confirmDeclareLostModal}
+        onConfirm={onConfirm}
         onCancel={this.onCancel}
         confirmLabel={<FormattedMessage id="ui-checkin.statusModal.confirm" />}
       />
+    );
+  }
+
+  renderWithdrawnModal() {
+    const {
+      showWithdrawnModal,
+      checkedinItem,
+    } = this.state;
+    const messageId = checkedinItem.discoverySuppress ?
+      'ui-checkin.withdrawnModal.suppressedMessage' :
+      'ui-checkin.withdrawnModal.notSuppressedMessage';
+
+    return this.renderConfirmStatusModal(
+      showWithdrawnModal,
+      'withdrawn',
+      messageId,
+      this.confirmWithdrawnModal,
+    );
+  }
+
+  renderDeclaredLostModal() {
+    const { showDeclaredLostModal } = this.state;
+    const messageId = 'ui-checkin.declaredLostModal.message';
+
+    return this.renderConfirmStatusModal(
+      showDeclaredLostModal,
+      'declaredLost',
+      messageId,
+      this.confirmDeclareLostModal,
     );
   }
 
@@ -309,18 +387,22 @@ class ModalManager extends React.Component {
 
   render() {
     const {
+      showClaimedReturnedModal,
       showMissingModal,
       showCheckinNoteModal,
       showMultipieceModal,
       showDeclaredLostModal,
+      showWithdrawnModal
     } = this.state;
 
     return (
       <>
+        {showClaimedReturnedModal && this.renderClaimedReturnedModal()}
         {showMissingModal && this.renderMissingModal()}
         {showCheckinNoteModal && this.renderCheckinNoteModal()}
         {showMultipieceModal && this.renderMultipieceModal()}
         {showDeclaredLostModal && this.renderDeclaredLostModal()}
+        {showWithdrawnModal && this.renderWithdrawnModal()}
       </>
     );
   }