+++ src/Scan.js
@@ -6,16 +6,10 @@ import {
   injectIntl,
 } from 'react-intl';
 import {
-  change,
-  reset,
-  SubmissionError,
-} from 'redux-form';
-import {
   get,
   isEmpty,
   keyBy,
   upperFirst,
-  uniqBy,
 } from 'lodash';
 
 import {
@@ -59,7 +53,10 @@ class Scan extends React.Component {
       }),
       items: PropTypes.shape({
         records: PropTypes.arrayOf(PropTypes.object),
-      })
+      }),
+      checkinSettings: PropTypes.shape({
+        records: PropTypes.arrayOf(PropTypes.object),
+      }),
     }),
 
     mutator: PropTypes.shape({
@@ -151,20 +148,29 @@ class Scan extends React.Component {
       path: 'circulation/end-patron-action-session',
       fetch: false,
     },
+    checkinSettings: {
+      type: 'okapi',
+      records: 'configs',
+      path: 'configurations/entries?query=(module=CHECKOUT and configName=other_settings)',
+    },
   });
 
   constructor(props) {
     super(props);
 
     this.state = {
-      loading: false
+      loading: false,
+      // itemClaimedReturnedResolution is a required checkin field for, unsurprisingly,
+      // items with status 'Claimed returned'. It is set in ClaimedReturnedModal via
+      // ModalManager.
+      itemClaimedReturnedResolution: null,
     };
   }
 
   store = this.props.stripes.store;
   barcode = React.createRef();
   checkInData = null;
-  checkinInst = null;
+  checkinFormRef = React.createRef();
   checkinInitialValues = {
     item: {
       checkinDate: '',
@@ -197,7 +203,7 @@ class Scan extends React.Component {
     }, []);
 
     this.clearResources();
-    this.clearForm('CheckIn');
+    this.clearForm();
 
     if (!isEmpty(uniquePatrons)) {
       const endSessions = uniquePatrons.map(patronId => ({
@@ -209,8 +215,8 @@ class Scan extends React.Component {
     }
   };
 
-  clearForm(formName) {
-    this.store.dispatch(reset(formName));
+  clearForm() {
+    this.checkinFormRef.current.reset();
   }
 
   clearResources() {
@@ -219,32 +225,45 @@ class Scan extends React.Component {
 
   validate(item) {
     const { intl: { formatMessage } } = this.props;
-    const barcode = formatMessage({ id: 'ui-checkin.fillOut' });
+    const checkin = formatMessage({ id: 'ui-checkin.fillOut' });
     if (!item || !item.barcode) {
-      throw new SubmissionError({ item: { barcode } });
+      return { checkin };
     }
+
+    return {};
   }
 
   onCloseErrorModal = () => {
     this.setState({ itemError: false },
       () => {
-        this.clearField('CheckIn', 'item.barcode');
+        this.clearField('item.barcode');
         this.setFocusInput();
       });
   }
 
-  tryCheckIn = async (data, checkInInst) => {
+  tryCheckIn = async (data) => {
+    const submitErrors = {};
     this.checkInData = data;
-    this.checkInInst = checkInInst;
-    this.validate(data.item);
+    const errors = this.validate(data.item);
+
+    if (!isEmpty(errors)) {
+      return errors;
+    }
+
     const { item: { barcode } } = data;
     const checkedinItem = await this.fetchItem(barcode);
 
     if (!checkedinItem) {
-      this.checkIn();
+      try {
+        await this.checkIn();
+      } catch (error) {
+        submitErrors.checkin = error;
+      }
     } else {
       this.setState({ checkedinItem });
     }
+
+    return submitErrors;
   }
 
   checkIn = () => {
@@ -262,6 +281,7 @@ class Scan extends React.Component {
       mutator: { checkIn },
       stripes: { user },
     } = this.props;
+    const { itemClaimedReturnedResolution } = this.state;
 
     const servicePointId = get(user, 'user.curServicePoint.id', '');
     const checkInDate = buildDateTime(checkinDate, checkinTime);
@@ -271,13 +291,19 @@ class Scan extends React.Component {
       itemBarcode: barcode.trim(),
     };
 
+    // For items that have the status 'Claimed returned', the claimedReturnedResolution
+    // parameter is required for checkin. For any other status, we don't want it.
+    if (itemClaimedReturnedResolution) {
+      requestData.claimedReturnedResolution = itemClaimedReturnedResolution;
+    }
+
     this.setState({ loading: true });
 
     return checkIn.POST(requestData)
       .then(checkinResp => this.processResponse(checkinResp))
       .then(checkinResp => this.fetchRequests(checkinResp))
       .then(checkinResp => this.addScannedItem(checkinResp))
-      .then(() => this.clearField('CheckIn', 'item.barcode'))
+      .then(() => this.clearField('item.barcode'))
       .catch(resp => this.processError(resp))
       .finally(() => this.processCheckInDone());
   }
@@ -285,7 +311,7 @@ class Scan extends React.Component {
   processResponse(checkinResp) {
     const { loan, item, staffSlipContext } = checkinResp;
     const checkinRespItem = loan || { item };
-    this.setState({ staffSlipContext });
+    this.setState({ staffSlipContext, itemClaimedReturnedResolution: null });
     if (get(checkinRespItem, 'item.status.name') === statuses.IN_TRANSIT) {
       checkinResp.transitItem = checkinRespItem;
       this.setState({ transitItem: checkinRespItem });
@@ -316,8 +342,7 @@ class Scan extends React.Component {
   }
 
   handleTextError(error) {
-    const item = { barcode: error };
-    throw new SubmissionError({ item });
+    throw error;
   }
 
   handleJsonError({
@@ -394,8 +419,8 @@ class Scan extends React.Component {
     return mutator.scannedItems.replace(scannedItems);
   }
 
-  clearField(formName, fieldName) {
-    this.props.stripes.store.dispatch(change(formName, fieldName, ''));
+  clearField = (fieldName) => {
+    this.checkinFormRef.current.change(fieldName, '');
   }
 
   throwError(error) {
@@ -435,6 +460,12 @@ class Scan extends React.Component {
     const spSlip = servicePoint.staffSlips.find(slip => slip.id === staffSlip.id);
 
     return (!spSlip || spSlip.printByDefault);
+  }
+
+  // Used by ModalManager and ClaimedReturnedModal to assign a claimedReturnedResolution
+  // value for items with the 'claimed returned' status, required for checkin.
+  claimedReturnedHandler = (resolution) => {
+    this.setState({ itemClaimedReturnedResolution: resolution });
   }
 
   renderHoldModal(request, staffSlipContext) {
@@ -603,7 +634,7 @@ class Scan extends React.Component {
   }
 
   onCancel = () => {
-    this.clearForm('CheckIn');
+    this.clearForm();
   };
 
   showCheckinNotes = (loan) => {
@@ -637,6 +668,7 @@ class Scan extends React.Component {
           <ModalManager
             checkedinItem={checkedinItem}
             checkinNotesMode={checkinNotesMode}
+            claimedReturnedHandler={this.claimedReturnedHandler}
             onDone={this.checkIn}
             onCancel={this.onCancel}
           />}
@@ -649,10 +681,11 @@ class Scan extends React.Component {
           loading={loading}
           scannedItems={scannedItems}
           items={items}
+          formRef={this.checkinFormRef}
           barcodeRef={this.barcode}
           initialValues={this.checkinInitialValues}
           showCheckinNotes={this.showCheckinNotes}
-          submithandler={this.tryCheckIn}
+          onSubmit={this.tryCheckIn}
           onSessionEnd={this.onSessionEnd}
           {...this.props}
         />