+++ src/consts.js
@@ -3,8 +3,10 @@ const statuses = {
   AWAITING_PICKUP: 'Awaiting pickup',
   MISSING: 'Missing',
   CHECK_IN: 'Check in',
+  CLAIMED_RETURNED: 'Claimed returned',
   AWAITING_DELIVERY: 'Awaiting delivery',
   DECLARED_LOST: 'Declared lost',
+  WITHDRAWN: 'Withdrawn',
 };
 
 const requestTypes = {
@@ -12,7 +14,13 @@ const requestTypes = {
   DELIVERY: 'Delivery',
 };
 
+const claimedReturnedResolutions = {
+  FOUND: 'Found by library',
+  RETURNED: 'Returned by patron',
+};
+
 export {
+  claimedReturnedResolutions,
   statuses,
   requestTypes,
 };