+++ CHANGELOG.md
@@ -8,7 +8,11 @@
 
 * Check for `accounts.collection.get` permission before rendering fee/fines components. Fixes UICHKIN-174.
 * Fix import path to stripes util. Fixes UICHKIN-175.
+* Show confirmation modal when item with withdrawn status is checked in. Refs UICHKIN-126.
 * Purge `intlShape` in prep for `react-intl` `v4` migration. Refs STRIPES-672.
+* Clear Check In Page When Session Expires. Refs UICHCKIN-177.
+* Add support for checking in Claimed returned items. Refs UICHKIN-116.
+* Add link to request details in action menu. Refs UICHKIN-103.
 
 ## [2.0.0] (https://github.com/folio-org/ui-checkin/tree/v2.0.0) (2019-03-13)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v1.10.0...v2.0.0)